Half-done logic to separate out contrib and non-contrib.
